requirements.txt
django-versatileimagefield>=1.10,<2.0
django-geostore==0.3.1
djangorestframework-jwt>=1.11,<1.12
django-cors-headers>=2.2,<2.3
django-url-filter>=0.3,<0.4
django-storages>=1.7,<1.8
boto3>=1.9,<=1.10
weasyprint>=44
simpleeval>=0.9
docxtpl>=0.5
python-magic>=0.4

# production
gunicorn
